Move around multi file reader files
