+++ extension/json/include/json_multi_file_info.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "duckdb/common/multi_file_reader_function.hpp"
+#include "duckdb/common/multi_file/multi_file_reader_function.hpp"
 #include "json_reader_options.hpp"
 
 namespace duckdb {
