+++ extension/json/include/json_scan.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "json_reader.hpp"
-#include "duckdb/common/multi_file_reader.hpp"
+#include "duckdb/common/multi_file/multi_file_reader.hpp"
 #include "duckdb/common/mutex.hpp"
 #include "duckdb/common/pair.hpp"
 #include "duckdb/common/types/type_map.hpp"
