+++ extension/parquet/parquet_metadata.cpp
@@ -5,7 +5,7 @@
 #include <sstream>
 
 #ifndef DUCKDB_AMALGAMATION
-#include "duckdb/common/multi_file_reader.hpp"
+#include "duckdb/common/multi_file/multi_file_reader.hpp"
 #include "duckdb/common/types/blob.hpp"
 #include "duckdb/common/types/column/column_data_collection.hpp"
 #include "duckdb/planner/filter/constant_filter.hpp"
