+++ src/common/CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(arrow)
 add_subdirectory(crypto)
 add_subdirectory(enums)
 add_subdirectory(exception)
+add_subdirectory(multi_file)
 add_subdirectory(operator)
 add_subdirectory(progress_bar)
 add_subdirectory(tree_renderer)
@@ -50,9 +51,6 @@ add_library_unity(
   http_util.cpp
   pipe_file_system.cpp
   local_file_system.cpp
-  multi_file_list.cpp
-  multi_file_reader.cpp
-  multi_file_reader_column_mapper.cpp
   error_data.cpp
   opener_file_system.cpp
   printer.cpp
