+++ src/include/duckdb/common/multi_file/base_file_reader.hpp
@@ -1,0 +1,121 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/common/multi_file/base_file_reader.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/common.hpp"
+#include "duckdb/common/types/value.hpp"
+#include "duckdb/common/column_index.hpp"
+#include "duckdb/planner/table_filter.hpp"
+#include "duckdb/common/optional_idx.hpp"
+#include "duckdb/common/multi_file/multi_file_reader_data.hpp"
+#include "duckdb/planner/expression.hpp"
+
+namespace duckdb {
+
+//! Parent class of single-file readers - this must be inherited from for readers implementing the MultiFileReader
+//! interface
+class BaseFileReader {
+public:
+	explicit BaseFileReader(string file_name_p) : file_name(std::move(file_name_p)) {
+	}
+	DUCKDB_API virtual ~BaseFileReader() = default;
+
+	//! The name of the file we are reading
+	string file_name;
+	//! (Optionally) The file index (generated by the multi file reader)
+	optional_idx file_list_idx;
+	//! The set of columns for the current file
+	vector<MultiFileReaderColumnDefinition> columns;
+	//! The column ids to read from the file
+	MultiFileLocalColumnIds<MultiFileLocalColumnId> column_ids;
+	//! The column indexes to read from the file
+	vector<ColumnIndex> column_indexes;
+	//! The set of table filters (adjusted to local indexes)
+	unique_ptr<TableFilterSet> filters;
+	//! Expression to execute for a given column (BEFORE executing the filter)
+	//! NOTE: this is only set when we have filters - it can be ignored for readers that don't have filter pushdown
+	unordered_map<column_t, unique_ptr<Expression>> expression_map;
+	//! The final types for various expressions - this is ONLY used if UseCastMap() is explicitly enabled
+	unordered_map<column_t, LogicalType> cast_map;
+
+public:
+	const vector<MultiFileReaderColumnDefinition> &GetColumns() const {
+		return columns;
+	}
+	const string &GetFileName() {
+		return file_name;
+	}
+	virtual string GetReaderType() const = 0;
+	virtual bool UseCastMap() const {
+		//! Whether or not to push casts into the cast map
+		return false;
+	}
+
+public:
+	template <class TARGET>
+	TARGET &Cast() {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<TARGET &>(*this);
+	}
+	template <class TARGET>
+	const TARGET &Cast() const {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<const TARGET &>(*this);
+	}
+};
+
+//! Parent class of file reader options
+class BaseFileReaderOptions {
+public:
+	DUCKDB_API virtual ~BaseFileReaderOptions() = default;
+
+public:
+	template <class TARGET>
+	TARGET &Cast() {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<TARGET &>(*this);
+	}
+	template <class TARGET>
+	const TARGET &Cast() const {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<const TARGET &>(*this);
+	}
+};
+
+//! Parent class of union data - used for the UNION BY NAME. This is effectively a cache that is kept around per file
+//! that can be used to speed up opening the same file again afterwards - to avoid doing double work.
+class BaseUnionData {
+public:
+	explicit BaseUnionData(string file_name_p) : file_name(std::move(file_name_p)) {
+	}
+	DUCKDB_API virtual ~BaseUnionData() = default;
+
+	string file_name;
+	shared_ptr<BaseFileReader> reader;
+	vector<string> names;
+	vector<LogicalType> types;
+
+	const string &GetFileName() {
+		return file_name;
+	}
+
+public:
+	template <class TARGET>
+	TARGET &Cast() {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<TARGET &>(*this);
+	}
+	template <class TARGET>
+	const TARGET &Cast() const {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<const TARGET &>(*this);
+	}
+};
+
+} // namespace duckdb
