+++ src/include/duckdb/common/multi_file/multi_file_reader_column_mapper.hpp
@@ -1,14 +1,14 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/common/multi_file_reader_column_mapper.hpp
+// duckdb/common/multi_file/multi_file_reader_column_mapper.hpp
 //
 //
 //===----------------------------------------------------------------------===//
 
 #pragma once
 
-#include "duckdb/common/multi_file_reader.hpp"
+#include "duckdb/common/multi_file/multi_file_reader.hpp"
 
 namespace duckdb {
 struct ResultColumnMapping;
