+++ src/include/duckdb/common/multi_file/multi_file_reader_data.hpp
@@ -1,22 +1,32 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/common/base_file_reader.hpp
+// duckdb/common/multi_file/multi_file_reader_data.hpp
 //
 //
 //===----------------------------------------------------------------------===//
 
 #pragma once
 
-#include "duckdb/common/common.hpp"
-#include "duckdb/common/types/value.hpp"
-#include "duckdb/common/column_index.hpp"
-#include "duckdb/planner/table_filter.hpp"
-#include "duckdb/common/optional_idx.hpp"
+#include "duckdb/function/table_function.hpp"
+#include "duckdb/function/copy_function.hpp"
+#include "duckdb/common/exception/conversion_exception.hpp"
 #include "duckdb/parser/expression/constant_expression.hpp"
-#include "duckdb/planner/expression.hpp"
+#include <numeric>
 
 namespace duckdb {
+
+enum class MultiFileFileState : uint8_t { UNOPENED, OPENING, OPEN, SKIPPED, CLOSED };
+
+struct HivePartitioningIndex {
+	HivePartitioningIndex(string value, idx_t index);
+
+	string value;
+	idx_t index;
+
+	DUCKDB_API void Serialize(Serializer &serializer) const;
+	DUCKDB_API static HivePartitioningIndex Deserialize(Deserializer &deserializer);
+};
 
 struct MultiFileReaderColumnDefinition {
 public:
@@ -278,104 +288,38 @@ private:
 	vector<MultiFileFilterEntry> filter_map;
 };
 
-//! Parent class of single-file readers - this must be inherited from for readers implementing the MultiFileReader
-//! interface
-class BaseFileReader {
+struct MultiFileConstantMap {
 public:
-	explicit BaseFileReader(string file_name_p) : file_name(std::move(file_name_p)) {
-	}
-	DUCKDB_API virtual ~BaseFileReader() = default;
-
-	//! The name of the file we are reading
-	string file_name;
-	//! (Optionally) The file index (generated by the multi file reader)
-	optional_idx file_list_idx;
-	//! The set of columns for the current file
-	vector<MultiFileReaderColumnDefinition> columns;
-	//! The column ids to read from the file
-	MultiFileLocalColumnIds<MultiFileLocalColumnId> column_ids;
-	//! The column indexes to read from the file
-	vector<ColumnIndex> column_indexes;
-	//! The set of table filters (adjusted to local indexes)
-	unique_ptr<TableFilterSet> filters;
-	//! Expression to execute for a given column (BEFORE executing the filter)
-	//! NOTE: this is only set when we have filters - it can be ignored for readers that don't have filter pushdown
-	unordered_map<column_t, unique_ptr<Expression>> expression_map;
-	//! The final types for various expressions - this is ONLY used if UseCastMap() is explicitly enabled
-	unordered_map<column_t, LogicalType> cast_map;
+	using iterator = vector<MultiFileConstantEntry>::iterator;
+	using const_iterator = vector<MultiFileConstantEntry>::const_iterator;
 
 public:
-	const vector<MultiFileReaderColumnDefinition> &GetColumns() const {
-		return columns;
+	template <typename... Args>
+	void Add(Args &&...args) {
+		constant_map.emplace_back(std::forward<Args>(args)...);
 	}
-	const string &GetFileName() {
-		return file_name;
+	const MultiFileConstantEntry &operator[](MultiFileConstantMapIndex constant_index) {
+		return constant_map[constant_index.index];
 	}
-	virtual string GetReaderType() const = 0;
-	virtual bool UseCastMap() const {
-		//! Whether or not to push casts into the cast map
-		return false;
+	idx_t size() const { // NOLINT: matching name of std
+		return constant_map.size();
 	}
-
-public:
-	template <class TARGET>
-	TARGET &Cast() {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<TARGET &>(*this);
+	// Iterator support
+	iterator begin() {
+		return constant_map.begin();
 	}
-	template <class TARGET>
-	const TARGET &Cast() const {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<const TARGET &>(*this);
+	iterator end() {
+		return constant_map.end();
 	}
-};
-
-//! Parent class of file reader options
-class BaseFileReaderOptions {
-public:
-	DUCKDB_API virtual ~BaseFileReaderOptions() = default;
-
-public:
-	template <class TARGET>
-	TARGET &Cast() {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<TARGET &>(*this);
+	const_iterator begin() const {
+		return constant_map.begin();
 	}
-	template <class TARGET>
-	const TARGET &Cast() const {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<const TARGET &>(*this);
-	}
-};
-
-//! Parent class of union data - used for the UNION BY NAME. This is effectively a cache that is kept around per file
-//! that can be used to speed up opening the same file again afterwards - to avoid doing double work.
-class BaseUnionData {
-public:
-	explicit BaseUnionData(string file_name_p) : file_name(std::move(file_name_p)) {
-	}
-	DUCKDB_API virtual ~BaseUnionData() = default;
-
-	string file_name;
-	shared_ptr<BaseFileReader> reader;
-	vector<string> names;
-	vector<LogicalType> types;
-
-	const string &GetFileName() {
-		return file_name;
+	const_iterator end() const {
+		return constant_map.end();
 	}
 
-public:
-	template <class TARGET>
-	TARGET &Cast() {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<TARGET &>(*this);
-	}
-	template <class TARGET>
-	const TARGET &Cast() const {
-		DynamicCastCheck<TARGET>(this);
-		return reinterpret_cast<const TARGET &>(*this);
-	}
+private:
+	vector<MultiFileConstantEntry> constant_map;
 };
 
 } // namespace duckdb
