+++ src/include/duckdb/execution/operator/csv_scanner/scanner_boundary.hpp
@@ -13,7 +13,7 @@
 #include "duckdb/execution/operator/csv_scanner/csv_reader_options.hpp"
 #include "duckdb/execution/operator/csv_scanner/csv_state_machine.hpp"
 #include "duckdb/main/client_context.hpp"
-#include "duckdb/common/multi_file_reader.hpp"
+#include "duckdb/common/multi_file/multi_file_reader.hpp"
 #include "duckdb/common/fast_mem.hpp"
 
 //! We all need boundaries every now and then, CSV Scans also need them
